fixed Saves, Castling Deleting Pieces
Castling is no longer allowed if there are pieces between the king and the rook.

Got getSaveMoves to behave normally for the time being.

Illegal moves that allow your king to remain in check are still in the game, as calculating those illegal moves was attempted, but I had to focus on other parts of the code.